Refactoring VPC provider
